# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.previewMergeResponse:
    properties:
      current:
        description: |-
          Current is the existing global values file body, exactly as returned
          by the GET values-schema endpoint. Returned so the UI can render a
          before/after diff without a second request.
        type: string
      diff_summary:
        allOf:
        - $ref: '#/definitions/internal_api.previewMergeSummary'
        description: |-
          DiffSummary is an at-a-glance view of what changed. Used by the modal
          header so the user can decide whether to open the diff or skip.
      merged:
        description: |-
          Merged is the candidate body the user can submit through the existing
          PUT /api/v1/addons/{name}/values endpoint. Same shape as Current
          (full YAML with smart-values header) plus any new upstream keys the
          user did not previously set.
        type: string
      upstream_version:
        description: |-
          UpstreamVersion is the chart version Sharko fetched the upstream
          values.yaml for. Echo of the catalog pin — surfaced so the UI can
          show "Merging in defaults from cert-manager@1.20.2".
        type: string
    type: object
  internal_api.previewMergeSummary:
    properties:
      new_keys:
        description: |-
          NewKeys are dotted paths present in upstream but NOT in the user's
          current file. These are the only paths that the merge actually adds.
        items:
          type: string
        type: array
      no_op:
        description: |-
          NoOp is true when the merge produced the same body as Current — the
          user already has every upstream key. The UI uses this to render a
          "Already up to date" empty state instead of an empty diff.
        type: boolean
      preserved_user_keys:
        description: |-
          PreservedUserKeys are dotted paths the user has set that already
          exist in upstream — the merge keeps the user's value and the user's
          inline comments are honored on the PR side. Surfaced so the modal
          can reassure the user nothing of theirs is being overwritten.
        items:
          type: string
        type: array
    type: object
...
  internal_api.setAddonValuesRequest:
    properties:
      refresh_from_upstream:
        description: |-
          RefreshFromUpstream, when true, regenerates the global values file
          from the chart's upstream values.yaml at the catalog-pinned version.
          The audit event for this path is `values_refreshed_from_upstream`
          (overriding the default `values_set`) so audit consumers can
          distinguish manual edits from upstream refreshes.
        type: boolean
      values:
        description: |-
          Values is the full YAML file content. Sharko commits this verbatim
          as the new global default values for the addon. Ignored when
          RefreshFromUpstream is true.
        type: string
    type: object
  internal_api.setClusterAddonValuesRequest:
    properties:
      values:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Addon management server for Kubernetes clusters, built on ArgoCD.
  title: Sharko API
  version: 1.0.0
paths:
...
  /addons/{name}/values:
    get:
      description: Returns the default Helm values file for a specific addon
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Addon values
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get addon values
      tags:
      - addons
    put:
      consumes:
      - application/json
      description: 'Replaces the global default Helm values YAML for an addon and
        opens a PR. Tier 2 (configuration) — prefers the caller''s personal GitHub
        PAT for proper Git authorship; otherwise falls back to the service token with
        a `Co-authored-by:` trailer and surfaces `attribution_warning: "no_per_user_pat"`
        in the response.'
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Full YAML payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.setAddonValuesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR created (or merged if auto-merge is on)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid YAML or missing field
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Addon not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git or ArgoCD failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Edit global addon values
      tags:
      - addons
...
  /addons/{name}/values/preview-merge:
    post:
      consumes:
      - application/json
      description: Returns a candidate values YAML that adds NEW upstream keys (and
        their default values) on top of the user's current file, preserving every
        key the user has already set. The endpoint is Tier 1 read-only — it does NOT
        open a PR. Submit the returned `merged` body through PUT /api/v1/addons/{name}/values
        to commit.
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.previewMergeResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Addon not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git or Helm-repo failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Preview a safe merge of upstream values into an addon's values file
      tags:
      - addons
...
  /clusters/{cluster}/addons/{name}/values:
    get:
      description: Returns the YAML for a single addon's section in a cluster's overrides
        file, plus an optional JSON Schema. `current_overrides` is empty when no overrides
        are configured for this addon yet.
      parameters:
      - description: Cluster name
        in: path
        name: cluster
        required: true
        type: string
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Current overrides + optional schema
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get per-cluster addon overrides + schema
      tags:
      - clusters
    put:
      consumes:
      - application/json
      description: Replaces the per-cluster overrides YAML for one addon on one cluster
        and opens a PR. Pass an empty `values` string to remove the addon's overrides
        entirely. Tier 2 (configuration) — same attribution semantics as the global
        values endpoint.
      parameters:
      - description: Cluster name
        in: path
        name: cluster
        required: true
        type: string
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Addon-section YAML payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.setClusterAddonValuesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR created (or merged if auto-merge is on)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid YAML or missing field
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git or ArgoCD failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Edit per-cluster addon overrides
      tags:
      - clusters
...
  /clusters/{name}/values:
    get:
      description: Returns the Helm values file for a specific cluster
      parameters:
      - description: Cluster name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster values
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cluster values
      tags:
      - clusters
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"